+++ src/fst/misc.py
@@ -217,9 +217,8 @@ Mode = Literal[
     the scope of desired return, for example `Constant` will parse as an expression but fail if the expression
     is not a `Constant`. These overlap with the string specifiers to an extent but not all of them. For example
     `AST` type `ast.expr` is the same as passing `'expr'`. Not all string specified modes are can be matched, for
-    example `'arguments_lambda'`. Likewise `'exec'` and `'stmts'` specify the same parse mode, but not the same as
-    `Module` since that is used as a general purpose slice container. `Tuple` parse also allows parsing `Slice`s in the
-    `Tuple`.
+    example `'arguments_lambda'`. Likewise `'exec'` and `'stmts'` specify the same parse mode. `Tuple` parse also allows
+    parsing `Slice`s in the `Tuple` as well as otherwise invalid star notation `*not a`.
 """