+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ### Fixed
 
+- disallow Starred as a target to Delete
 - allow put to except and case slices
 - allow put any kind of expression slice to tuple, including Slices
 - parse functions handle trailing comments without newlines correctly
@@ -9,6 +10,7 @@
 
 ### Added
 
+- prescribed slicing for `Delete.targets`
 - prescribed slicing for `FunctionDef.type_params`
 - prescribed slicing for `AsyncFunctionDef.type_params`
 - prescribed slicing for `ClassDef.type_params`